+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Removed support for Ruby 2.1.*
 - Ruby 2.3.1 as default
 - Gem dump
+- Dry configurable config (#20)
 
 ## 0.3.1
 - Dev tools update